services

+++ src/main/java/com/example/condominio/Repositories/EdificioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EdificioRepository extends JpaRepository<Edificio, Long> {
+public interface EdificioRepository extends JpaRepository<Edificio, Integer> {
 }
